refactor: redux store & minor cleanup
