+++ src/layouts/MainLayout/TopBar/Account.js
@@ -13,10 +13,7 @@ import {
   CircularProgress,
   makeStyles,
 } from "@material-ui/core";
-import { dismissLogin, login, logout } from "../../../actions/accountActions";
-// import { cfaSignIn } from 'capacitor-firebase-auth';
-// import authService from '../../../services/authService';
-// import { useSnackbar } from 'notistack';
+import { dismissLogin, login, logout } from "../../../store/actions/accountActions";
 
 const useStyles = makeStyles((theme) => ({
   button: {
